split out QuiltParse
