+++ quiltplus/__init__.py
@@ -1,4 +1,5 @@
 from quiltplus.cache import *
 from quiltplus.id import *
 from quiltplus.package import *
+from quiltplus.parse import *
 from quiltplus.unparse import *
